Rename de problem 1 on 7-12-2022.txt to de problem 1 on 7-12-2022.c
1.	Write a high level program for the following:
Suppose that the data for analysis includes the attribute age. The age values for the data tuples are (in increasing order)
13, 15, 16, 16, 19, 20, 20, 21, 22, 22, 25, 25, 25, 25, 30, 33, 33, 35, 35, 35, 36, 40, 45, 52, 70.
a.	What is the mean of the data? What is median?
b.	What is the mode of the data? Comment on the data’s modality (bimodal, trimodal, etc..)
c.	What is mid-range of the data?
d.	Can you find the first quartile(Q1) and the third quartile (Q3) of the data?
e.	Give the five number summary of the data.